mycelo: Moves genesis commands to their own file
